+++ apps/user/src/app/signup/signup.hooks.ts
@@ -4,7 +4,6 @@ import {
   useVerificationMutation,
 } from '@/services/auth/mutations';
 import type { Join } from '@/types/auth/client';
-import { useBooleanState } from '@maru/hooks';
 import type { ChangeEventHandler } from 'react';
 import { useState } from 'react';
 
@@ -26,13 +25,10 @@ export const useJoinAction = (joinUserData: Join, termsAgree: boolean) => {
   return { handleJoin };
 };
 
-export const useVerificationAction = (joinUserData: Join) => {
-  const { value: isVerificationCodeSend, setValue: setIsVerificationCodeSend } =
-    useBooleanState(false);
-  const { value: isVerificationCodeDisabled, setValue: setisVerificationCodeDisabled } =
-    useBooleanState(false);
-  const { value: isVerificationCodeConfirm, setValue: setisVerificationCodeConfirm } =
-    useBooleanState(false);
+export const useVerificationCodeAction = (joinUserData: Join) => {
+  const [isVerificationCodeSend, setIsVerificationCodeSend] = useState(false);
+  const [isVerificationCodeDisabled, setisVerificationCodeDisabled] = useState(false);
+  const [isVerificationCodeConfirm, setisVerificationCodeConfirm] = useState(false);
 
   const { verificationMutate } = useVerificationMutation(setisVerificationCodeConfirm);
   const { requestVerificationCodeMutate } = useRequestVerificationCodeMutation(
